powerpc: Fix whitespace merge in mpc8641 hpcn device tree
When we coverted the .dts to v1 we lost a space between the irq
and its polarity/sense information.  This causes a bit of chaos
as the reset of the blob is off by one cell.

This was noticed by booting and getting errors w/ATA due to
lock of interrupts.

Signed-off-by: Kumar Gala <galak@kernel.crashing.org>

[ upstream commit: ba1616d921429ffe7480e8835e85f95ff041add8 ]