ARM: ux500: Change IRQ from low-to-high edge triggered to high-to-low
When the STMPE IRQ is triggered to be active high level-sensitive, the
Nomadik GPIO controller it uses complains, although it still works.
Recently we attempted to move triggering to low-to-high in an attempt
to prevent the warning; however, this ensured that the IRQ was actually
missed completely. Now we have a solution which both works and keeps the
GPIO controller happy.

Signed-off-by: Lee Jones <lee.jones@linaro.org>
Signed-off-by: Linus Walleij <linus.walleij@linaro.org>

[ upstream commit: 7cf261dd1efc593942db913c98851feadec9d73e ]