ARM: bcm2835: add SPI device to DT
The BCM2835 has a single instance of the "SPI0"-type SPI master
controller. Instantiate it in the SoC .dtsi file, Don't enable it in
the Raspberry Pi board .dts file, since we have no idea what is actually
connected, and hence no idea what to set the bus clock rate to.

Signed-off-by: Stephen Warren <swarren@wwwdotorg.org>

[ upstream commit: 6ce5f02ef7edd33ebdac27953b52c489a6005e6d ]