ARM: ux500: Supply UART's DMA configuration via Device Tree
When requesting a channel, a DMA client needs to pass some pieces of
information such as; request channel, device type, channel type and
direction etc. Normally we do this in the form of platform data, but
when DT is enabled we need to pass it using the driver's bindings
instead.

Signed-off-by: Lee Jones <lee.jones@linaro.org>
Signed-off-by: Linus Walleij <linus.walleij@linaro.org>

[ upstream commit: fbff01cce09fb79a04619a9da782db655fbd7df0 ]