ARM: nomadik: switch to use the Nomadik I2C driver
Instead of using bit-banged I2C, let's use the actual I2C
driver in the kernel. Since the I2C block may be communicating
with things like the PMIC, we need to select it from the Kconfig
just like the bit-banged adapter is selected today. The rest of
the configuration for this driver can be done from the device
tree.

Signed-off-by: Linus Walleij <linus.walleij@linaro.org>

[ upstream commit: 66e0c12f9e17a1c97fc358f9e1f5e1dc615e279a ]