ARM: dts: imx51-babbage: Make DVI and WVGA panel functional
Currently we get the following errors on imx51-babbage:

/display@di0: could not find display-timings node
/display@di0: no timings specified
/display@di1: could not find display-timings node
/display@di1: no timings specified
imx-drm imx-drm: failed to allocate buffer with size 0

Provide timing values for IPU1, which is connected to a DVI bridge and
to IPU2, which can be connected to the WVGA panel, so that both of them can
be functional.

While at it, disable the WVGA panel, so that DVI output becomes the default one.

Signed-off-by: Fabio Estevam <fabio.estevam@freescale.com>
Signed-off-by: Shawn Guo <shawn.guo@linaro.org>

[ upstream commit: 493a86365bebdae944a9f2bbf21e106a77c92928 ]