ARM: dts: Add common support for omap3-evm
Looks like the main difference between the TMDSEVM3530 and
TMDSEVM3730 is just the omap processor:

http://www.ti.com/tool/tmdsevm3530
http://www.ti.com/tool/tmdsevm3730

So let's add a common file for the EVMs, and fix the description
for the omap3-evm.dst as that's clearly for the TMDSEVM3530
since it includes omap34xx.dtsi. It cannot support the TMDSEVM3730
properly, and we need a separate file for that in the following
patch.

Signed-off-by: Tony Lindgren <tony@atomide.com>

[ upstream commit: 465ce68b467e5e47d699564432dff8cb60204433 ]