ARM: nomadik: get rid of explicit ethernet GPIO management
We used to set up the GPIO used for the SMC ethernet IRQ by
requesting the GPIO line explicitly. As we recently established
that specifying the GPIO controller as interrupt parent and
requesting one of its interrupts by reference should be enough,
we do away with this. Incidentally, the Nomadik GPIO chip driver
handles this just fine: the IRQ is set up and fired as expected.

Signed-off-by: Linus Walleij <linus.walleij@linaro.org>

[ upstream commit: 6568eccd10f6d417f7e40160ca29c965a34b6254 ]