ARM: sun6i: colombus: Enable the I2C controllers
The A31 Colombus board has 3 I2C controllers that should be usable. However,
the first one is not working for some reason on the hardware I have been able
to test it on, while it should really be the same controller. Enable the i2c1
and i2c2 busses, and mark i2c0 as in failure in the DT.

Signed-off-by: Maxime Ripard <maxime.ripard@free-electrons.com>

[ upstream commit: 1cff74027cbd05d8ab8adf24e74ac7b4db1b2011 ]