Documentation: dt: iio: improve atmel-adc documentation
This corrects the example by adding #address-cells, #size-cells and a reg for
each trigger to comply to the ePAPR, as suggested by Mark Rutland in
http://lists.infradead.org/pipermail/linux-arm-kernel/2014-February/234184.html

Also, it removes the properties that are not used anymore, to stop propagating
them.
Also fixes atmel,adc-use-external-triggers property name.

Finally, fixes a few typos.

Signed-off-by: Alexandre Belloni <alexandre.belloni@free-electrons.com>
Signed-off-by: Nicolas Ferre <nicolas.ferre@atmel.com>

[ upstream commit: b5303ffbcf1194cbd5e509579ee7e28c629247b7 ]