ARM: dts: bcm28155-ap: leave camldo1 on to fix reboot
The BCM28155-AP board has a bootloader that expects the camldo1
regulator to be enabled on entry. Currently, the camldo1 regulator
is disabled when no longer in use as is the case during a reboot /
warm reset. This causes the early bootloader to hang upon entry. Add
regulator-always-on to the camldo1 constraint to fix reboot.

Reported-by: Alex Elder <elder@linaro.org>
Signed-off-by: Matt Porter <mporter@linaro.org>
Tested-by: Alex Elder <elder@linaro.org>

[ upstream commit: a97d67d1c4c331644f26704976f92a5b160c3f77 ]