ARM: dts: rockchip: enable vops and hdmi output on rk3288-firefly and -evb
This enables the recently added display nodes on the firefly and evb,
letting the boards output graphic data over the hdmi port.

This includes enabling of i2c5 on the evb that is used for ddc probing and
setting the lcd-regulators to always on until the hdmi driver learns
about using them.

Signed-off-by: Heiko Stuebner <heiko@sntech.de>

[ upstream commit: 863b76b2b5da45cdf8819c4db468adfd2f0292f0 ]