ARM: tegra: Add EMC timings to Nyan Blaze device tree
This adds a new file, tegra124-nyan-blaze-emc.dtsi that contains
valid timings for the EMC memory clock. The file is included to the
main device tree file for the Nyan Blaze.

The frequency 528MHz is missing because we don't currently have a timing
configuration that works.

Additionally, only the timings for the ram-code 1 is present as that's
what could be tested currently, though downstream has timings for more
ram-codes.

Signed-off-by: Tomeu Vizoso <tomeu.vizoso@collabora.com>
Signed-off-by: Thierry Reding <treding@nvidia.com>

[ upstream commit: 03650bd2010b117c5c16a75278b0d01def8e9376 ]