ARM: dts: sun8i: Enable USB DRC on Ippo Q8H-A23 tablets
Enable the otg controller on Ippo Q8H tablets, for now it is enabled
in host-only mode, because true OTG support requires support for
detecting and enabling Vbus through the axp221 pmic.

For this to work the Vbus on the port must be enabled by u-boot,
or a powered hub must be used.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>
Signed-off-by: Maxime Ripard <maxime.ripard@free-electrons.com>

[ upstream commit: a85b59c14d4082abcf27f63994e900dd00b5e572 ]