filter.sh: Rewrite upstream tags with -raw suffix
This allows them to be merged into master in this branch (which
contains the build system etc) and then tagged vX.Y-dts, so the tag
will actually be a useful tree.

Existing historival vX.Y-dts tags will be unchanged but will gain a
matching vX.Y-dts-raw, only new vX.Y-dts will be the merge.