ARM: dts: imx6*-hummingboard: fix pcie reset GPIO specification
PCIe reset signals are active low, and our GPIO for this is directly
connected to the PCIe reset.  However, as the PCIe driver was not using
the flag, the specification of '0' flags (which means active high) had
not been noticed.  Correct this oversight, and switch to using the
GPIO flag definitions instead.

Signed-off-by: Russell King <rmk+kernel@arm.linux.org.uk>
Signed-off-by: Shawn Guo <shawnguo@kernel.org>

[ upstream commit: 5cff48fc1e2b5efcb2e9f8d36a288dbd3d8df8b8 ]