PCI: rcar: Add gen2 device tree support for r8a7793
Add "renesas,pci-r8a7793" as a compatibility string for
"renesas,pci-rcar-gen2".

This doesn't change the driver, so it does nothing by itself.  But it
does mean that checkpatch won't complain about a future patch that adds
"renesas,pci-r8a7793" to a DT, which helps ensure that shipped DTs use
documented compatibility strings.

[bhelgaas: changelog]
Signed-off-by: Simon Horman <horms+renesas@verge.net.au>
Signed-off-by: Bjorn Helgaas <bhelgaas@google.com>
Acked-by: Rob Herring <robh@kernel.org>
[ upstream commit: a176cd304a8a07d6d9191126fd9bece4f67358c3 ]