ARM: dts: sun8i: Fix wrong Quad core / a33 compat for sun8i-a23-gt90h-v4
As the dts file name already implies sun8i-a23-gt90h-v4.dts is for
an a23 equipped tablet. I don't know how the "Quad Core" or a33
comaptible got in there, likely a copy and paste error.

Regardless this commit fixes this, note this is almost
purely a cosmetical fix, for all things that matter at the
machine compatible level the a23 and a33 are compatible.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>
Signed-off-by: Maxime Ripard <maxime.ripard@free-electrons.com>

[ upstream commit: 388bc9a70fddef3102992d70d45e57e5b3beb496 ]