ARM: dts: sun8i: Base Orange Pi Plus dts on the Orange Pi 2 dts
The Orange Pi Plus really is an Orange Pi 2 extended with:
1) A sata <-> usb bridge connected to ehci3
2) An eMMC on mmc2
3) An external gigabit ethernet phy instead of the integrated 100Mbit phy

This commit changes the dts to reflect this by making it include
the Orange Pi 2 dts and then adding the extra bits.

Note that the difference in ethernet phy is not taken into account
because we do not have an ethernet driver for the H3 yet.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>
Signed-off-by: Maxime Ripard <maxime.ripard@free-electrons.com>

[ upstream commit: 0c1747a24855f6ce9668283be0b6839f4c0ab4cf ]