ARM: dts: Fix uart wakeirq on omap5 by removing WAKEUP_EN for omaps
The padconf register WAKEUP_EN is now handled in a generic way using
Linux wakeirqs where pinctrl-single toggles the WAKEUP_EN bit when
a wakeirq is enabled or disabled.

At least omap5 gets confused if the WAKEUP_EN bit is set and the pin
is not claimed as a wakeirq. The end result is that wakeirqs don't
work properly as there is nothing handling the wakeirq.

So let's just remove the WAKEUP_EN usage from dts files.

Signed-off-by: Tony Lindgren <tony@atomide.com>

[ upstream commit: e0e80d43fc0c5aace76a6c5f9462a1aec3e004f8 ]