ARM: dts: rockchip: move rk3288 io-domain nodes to the grf
io-voltage control is actually part of the grf, so move the node under the
newly available grf simple-mfd.

To minimize duplicate code, the core node and compatible property
gets placed in the core rk3288.dtsi while the individual boards
now only need to enable it and add the necessary supply properties.

Signed-off-by: Heiko Stuebner <heiko@sntech.de>

[ upstream commit: 3445b2fae55f1a232e6e4843f4f74de074d8c07f ]