arm64: dts: marvell: add description for the Armada 8040 dev board
This commit adds a Device Tree description for the Marvell Armada 8040
Development Board. It features a quad-core Cortex A72 Armada 8040 SoC,
with a large number of peripherals: dual Gigabit, dual 10 GBit, 6 PCIe
interfaces, 6 SATA ports, 4 USB 3.0 ports, and more.

Only a subset of the functionalities are supported so far, and
additional features will be progressively enabled in the future.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>
Signed-off-by: Gregory CLEMENT <gregory.clement@free-electrons.com>

[ upstream commit: ec03445c9ee282eef12b72cdb029348f80daad58 ]