ARM: dts: exynos: Use stdout non-deprecated property and add serial options to Odroid XU/XU3/XU4
The "linux,stdout-path" is deprecated.

Add serial options so the console can be properly configured when
bootloader does not pass "console" parameter to the command line.  Without
options and lack of "console" in command line, the serial output dies
during boot (just after probing sound, probably because of rate change).

Signed-off-by: Krzysztof Kozlowski <k.kozlowski@samsung.com>
Reviewed-by: Javier Martinez Canillas <javier@osg.samsung.com>

[ upstream commit: 5a5232d1932b5d1ca472037bc8b38c21fa0cb6f8 ]