ARM: dts: armada-xp-matrix: Fix the location of the pcie-controller node
In the dts for the Marvell Armada XP Matrix board the pcie-controller was
located under the internal-regs node whereas it belongs to the soc node.

It means that, until this fix, the pcie could not work for this board
because it didn't match the definition of the pcie-controller node in the
dtsi file. If we had a look on the decompiled dtb file we saw two
different instances of the pcie-controller node: one with the all the
resource set but disabled and the other without any resource but enabled.

Signed-off-by: Gregory CLEMENT <gregory.clement@free-electrons.com>

[ upstream commit: 1a9cd8f36d0617630b4f0ae2601fe5600edff109 ]