ARM: dts: Ux500: move USB PHY pins to PHY device
The physical pins from the SoC are in a sense belonging to the
PHY device (AB8500 USB) rather than the MUSB USB IP block.
The driver definately assumes so: before this change it
complains that it cannot control the pins it is using:

abx5x0-usb ab8500-usb.0: could not get/set default pinstate

After this patch the warning goes away.

Cc: Ulf Hansson <ulf.hansson@linaro.org>
Signed-off-by: Linus Walleij <linus.walleij@linaro.org>

[ upstream commit: 6097c2d0afa4ff70ece8610875a0f0592064de29 ]