ARM: dts: imx: ventana: fix DTC warnings
Remove the sky2 ethernet device node from the pcie controller which was
invalid to begin with.

The original intent was to allow the bootloader to populate the MAC via
dt but this requires the PCI bus topology to be complete in dt as well
and as these boards have an expansion connector that topology is dynamic
and can't be represented here.

Signed-off-by: Tim Harvey <tharvey@gateworks.com>
Signed-off-by: Shawn Guo <shawnguo@kernel.org>

[ upstream commit: 1b1ec50366a87ed84ac07db9a09c56ac7cafb367 ]