ARM: sun8i: a83t: cubietruck-plus: Enable AC100 combo chip in AXP818 PMIC
The AXP813/AXP818 PMICs used with the A83T/H8 SoCs are actually 2 dies
in one package sharing the serial bus (I2C/RSB) pins. One die is the
actual PMIC. The other is an AC100 codec / RTC combo chip.

This patch adds the device nodes for the AC100 chip to the Cubietruck
Plus device tree.

Signed-off-by: Chen-Yu Tsai <wens@csie.org>
Acked-by: Maxime Ripard <maxime.ripard@free-electrons.com>

[ upstream commit: 0c62fb093e4093de1bc7432c4ffc2dca44d73c51 ]